Created base structure for window rendering

+++ deploy.sh
@@ -1,4 +1,7 @@
-rm -rf ./build/*
+
+if [[ $1 == "clean" ]]; then
+	rm -rf ./build/*
+fi
 cmake -S . -B ./build/
 cd ./build/; make;
 ./2DGame
